move templates to separate folder
