+++ compiler/docs/lona-workspace/components/ComponentTitle.component
@@ -1,0 +1,83 @@
+{
+  "devices" : [
+    {
+      "deviceId" : "iPhone SE",
+      "heightMode" : "At Least"
+    },
+    {
+      "deviceId" : "Pixel 2",
+      "heightMode" : "At Least"
+    }
+  ],
+  "examples" : [
+    {
+      "id" : "Default",
+      "name" : "Default",
+      "params" : {
+
+      }
+    }
+  ],
+  "logic" : [
+    {
+      "assignee" : [
+        "layers",
+        "ComponentTitle",
+        "text"
+      ],
+      "content" : [
+        "parameters",
+        "name"
+      ],
+      "type" : "AssignExpr"
+    },
+    {
+      "assignee" : [
+        "layers",
+        "ComponentIntro",
+        "text"
+      ],
+      "content" : [
+        "parameters",
+        "intro"
+      ],
+      "type" : "AssignExpr"
+    }
+  ],
+  "params" : [
+    {
+      "name" : "name",
+      "type" : "String"
+    },
+    {
+      "name" : "intro",
+      "type" : "String?"
+    }
+  ],
+  "root" : {
+    "children" : [
+      {
+        "id" : "ComponentTitle",
+        "params" : {
+          "font" : "h1",
+          "marginBottom" : 54,
+          "text" : "Component Title"
+        },
+        "type" : "Lona:Text"
+      },
+      {
+        "id" : "ComponentIntro",
+        "params" : {
+          "font" : "large",
+          "text" : "Text goes here"
+        },
+        "type" : "Lona:Text"
+      }
+    ],
+    "id" : "View",
+    "params" : {
+      "alignSelf" : "stretch"
+    },
+    "type" : "Lona:View"
+  }
+}
