+++ compiler/docs/templates/colors.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Layout from '../components/Layout'
+import Layout from '../src/components/Layout'
 
 export default function Template({ pageContext: { colors }, location }) {
   return (
