+++ compiler/docs/templates/textStyles.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Layout from '../components/Layout'
+import Layout from '../src/components/Layout'
 
 export default function Template({ pageContext: { textStyles }, location }) {
   return (
